+++ src/plugins/plot/spectralAggregatePlot/SpectralAggregatePlotConstants.js
